v1.04_1812/requirements.utf8.txt
altgraph==0.17.4
auto-py-to-exe==2.45.0
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.1
CTkMessagebox==2.7
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
Eel==0.18.1
et_xmlfile==2.0.0
fonttools==4.55.3
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
idna==3.10
kiwisolver==1.4.7
matplotlib==3.10.0
mysql-connector-python==9.1.0
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyMuPDF==1.25.1
pyparsing==3.2.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.6.0
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zope.event==5.0
zope.interface==7.2
